feat(picker): integrate picker-column-option with picker (#28666)
Issue number: N/a

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

The team would like to add a way of using the picker inline with page
content.

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- This PR integrates the column option component with the rest of the
picker so that is functional. This PR is a combination of commits from
previously reviewed PRs.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->

---------

Co-authored-by: ionitron <hi@ionicframework.com>