test(many): add/update axe tests for form controls (#28669)
Issue number: N/A

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

Axe tests are not run for light/dark theme

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Re-enabled the color-contrast test for range. This was disabled due to
a (now fixed) Axe bug.
- Moved the HTML templates into `setContent` so we can more easily test
light/dark mode
- Added tests for each component such that the color palette is used.
Checked states were added for checkbox, radio, and toggle. Focused
states were added for input, textarea, and select. The value of the
range was set to 50% so we can see more of the selected track.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->