test(overlays): add Axe tests for color contrast (#28630)
Issue number: Internal

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

- The action-sheet, alert, and toast components have tests for Axe
violations, but the `color-contrast` rule is disabled.
- The select-popover component does not test for Axe violations.
- For all of the above plus loading, dark mode test for color contrast
are not included.

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Existing Axe tests pulled out into separate `configs` blocks that test
both light and dark themes. Markup has also been converted to use
`page.setContent` instead of `page.goto` to support automatic theme
switching.
- `color-contrast` rule re-enabled on action-sheet, alert, and toast.
The rule was disabled due to contrast issues with the page content
covered up by the backdrop, but with the new setup, the overlay is the
only thing on the page.
- For select-popover, the Axe test has been limited to only color
contrast testing for now. The component had several Axe failures that
will be addressed separately to keep scope down. (We didn't previously
have any Axe testing on this component, which is why this wasn't
caught.)
- HTML file for the loading a11y test removed, since we're no longer
using it.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->