feat(alert): update styles to iOS 17 specs (#28726)
Issue number: N/A

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

Alerts don't match the iOS 17 spec for styling.

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Overlay border radius increased.
- Input font size increased.
- Input padding increased.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!--
  If this introduces a breaking change:
1. Describe the impact and migration path for existing applications
below.
  2. Update the BREAKING.md file with the breaking change.
3. Add "BREAKING CHANGE: [...]" to the commit description when merging.
See
https://github.com/ionic-team/ionic-framework/blob/main/.github/CONTRIBUTING.md#footer
for more information.
-->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->

The difference in the MD screenshots is the box shadow on the header in
the background. This was a change made a while back that probably wasn't
caught due to the pixel threshold on the screenshot update job. Deleting
all screenshots and running again got it updated.

---------

Co-authored-by: ionitron <hi@ionicframework.com>
Co-authored-by: Liam DeBeasi <liamdebeasi@users.noreply.github.com>