feat(themes): add dark theme files and mixins (#28767)
Issue number: Internal

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->
The dark theme is not exported for use in Ionic Framework apps

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- The dark theme is usable in Ionic Framework apps via any of the
following:
- The `ion-theme-dark` class (once the developer assigns this class to
the `html` element in their app) via dark.class.css
  - The `prefers-color-scheme` media query via dark.system.css
  - Throughout the full app via dark.always.css

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!--
  If this introduces a breaking change:
1. Describe the impact and migration path for existing applications
below.
  2. Update the BREAKING.md file with the breaking change.
3. Add "BREAKING CHANGE: [...]" to the commit description when merging.
See
https://github.com/ionic-team/ionic-framework/blob/main/.github/CONTRIBUTING.md#footer
for more information.
-->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->
Usage:

```
@import '~@ionic/core/css/themes/dark.always.css';
/* @import '~@ionic/core/css/themes/dark.class.css'; */
/* @import '~@ionic/core/css/themes/dark.system.css'; */
```

replace `~@ionic/core` with the relevant package

Dev build: `7.5.8-dev.11704294569.1abc989d`

Note to reviewers: the design doc says to include these files in
utils.bundle.scss. Doing this causes the dark theme to get applied by
default because it includes the "always" theme in the bundled css used
in Framework. Adding the imports in that file is not required for the
files to get included in the built package; users can import them
without them being included there.

We can bikeshed the name of the "system" file.